fix: refactoring personal details for simplicity
